move constants from edm folder
