+++ src/utils/constants/GlobalConstants.js
@@ -1,8 +1,6 @@
 // @flow
 
-
 const PARTICIPANTS_PREFIX = 'chronicle_participants_';
-
 const STUDY_LOGIN_ROOT = 'https://openlattice.com/chronicle/login?studyId=';
 
 export {
